Final_Project/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "edgar-analyst-copilot"
version = "0.2.0"
description = "Financial RAG pipeline with XBRL grounding and streaming chat"
requires-python = ">=3.10"
authors = [
  {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "requests>=2.32.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "faiss-cpu>=1.7.4",
    "tiktoken>=0.7.0",
    "openai>=1.35.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.7.0",
    "pyarrow>=15.0.0",
    "streamlit>=1.35.0"
]

[tool.setuptools.packages.find]
include = ["app*", "helper_lib*"]